Extract DeleteIndexCallable
Extract DeleteIndexCallable from QueryImpl into a standalone callable.
